+++ htdocs/database/migrations/2024_02_23_120137_create_race_drivers_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('race_drivers', function (Blueprint $table) {
             $table->id();
-            $table->integer('driverId');
-            $table->integer('raceId');
+            $table->unsignedBigInteger('driverId');
+            $table->unsignedBigInteger('raceId');
             $table->integer('dorsal')->nullable();
             $table->timestamp('time')->nullable();
